Move `I/JRecord` structs to dedicated files`
This makes it a bit easier to find them and brings the `extension` module size down a little
