+++ src/records/extension.rs
@@ -172,46 +172,12 @@ impl<'a> ExtensionDefRecord<'a> {
         })
     }
 
-    fn fmt(&self, f: &mut fmt::Formatter, letter: char) -> fmt::Result {
+    pub(crate) fn fmt(&self, f: &mut fmt::Formatter, letter: char) -> fmt::Result {
         write!(f, "{}{:02}", letter, self.num_extensions)?;
         for ext in self.extensions.iter() {
             write!(f, "{}", ext)?;
         }
         Ok(())
-    }
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct IRecord<'a>(pub ExtensionDefRecord<'a>);
-
-impl<'a> IRecord<'a> {
-    pub fn parse(line: &'a str) -> Result<Self, ParseError> {
-        let first_byte = line.as_bytes()[0];
-        assert!(first_byte == b'I');
-        Ok(IRecord(ExtensionDefRecord::parse(line)?))
-    }
-}
-
-impl<'a> fmt::Display for IRecord<'a> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        self.0.fmt(f, 'I')
-    }
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct JRecord<'a>(pub ExtensionDefRecord<'a>);
-
-impl<'a> JRecord<'a> {
-    pub fn parse(line: &'a str) -> Result<Self, ParseError> {
-        let first_byte = line.as_bytes()[0];
-        assert!(first_byte == b'J');
-        Ok(JRecord(ExtensionDefRecord::parse(line)?))
-    }
-}
-
-impl<'a> fmt::Display for JRecord<'a> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        self.0.fmt(f, 'J')
     }
 }
 
@@ -233,35 +199,5 @@ mod tests {
         };
 
         assert_eq!(parsed_record, expected);
-    }
-
-    #[test]
-    fn irecord_format() {
-        let expected_string = "I033638FXA3941ENL4246TAS";
-        let record = IRecord(ExtensionDefRecord {
-            num_extensions: 3,
-            extensions: vec![
-                Extension::new("FXA", 36, 38),
-                Extension::new("ENL", 39, 41),
-                Extension::new("TAS", 42, 46),
-            ],
-        });
-
-        assert_eq!(format!("{}", record), expected_string);
-    }
-
-    #[test]
-    fn jrecord_format() {
-        let expected_string = "J033638FXA3941ENL4246TAS";
-        let record = JRecord(ExtensionDefRecord {
-            num_extensions: 3,
-            extensions: vec![
-                Extension::new("FXA", 36, 38),
-                Extension::new("ENL", 39, 41),
-                Extension::new("TAS", 42, 46),
-            ],
-        });
-
-        assert_eq!(format!("{}", record), expected_string);
     }
 }
